+++ src/interactions/events/createMoonlit.ts
@@ -12,10 +12,10 @@ import {
     CacheType,
     CommandInteraction,
   } from 'discord.js';
-  import { getMoonlitTemplate } from '../utils/moonlitTemplate';
+  import { getMoonlitTemplate } from '../utils/templates/moonlitTemplate';
   dotenv.config();
 
-  export const handleCreateEvent = async (
+  export const handleMoonlitEvent = async (
     interaction: CommandInteraction<CacheType>,
   ) => {
     if (!interaction.isCommand()) return;
@@ -23,10 +23,10 @@ import {
     const options =
       interaction.options as CommandInteractionOptionResolver<CacheType>;
     const startTimeUnix = options.getString('starttime', true);
-    const slots = options.getInteger('slots', true);
-    const duration = options.getInteger('duration', true);
-    const dj = options.getString('DJ', true);
-    const leader = options.getString('Leader', true);
+    const slots = 4;
+    const duration = 15;
+    const dj = options.getString('dj', true);
+    const leader = options.getString('leader', true);
   
     // Generate slot timestamps
     const slotTimestamps = [];
@@ -50,7 +50,7 @@ import {
   
     const modal = new ModalBuilder()
       .setCustomId('createEventModal')
-      .setTitle('Create Event');
+      .setTitle('Create Moonlit Event');
   
     const templateInput = new TextInputBuilder()
       .setCustomId('template')
@@ -65,47 +65,3 @@ import {
   
     await interaction.showModal(modal);
   };
-  
-  export const handleCreateEventModalSubmit = async (
-    interaction: Interaction,
-  ) => {
-    if (!interaction.isModalSubmit()) return;
-  
-    const template = interaction.fields.getTextInputValue('template');
-  
-    // Create buttons
-    const signUpButton = new ButtonBuilder()
-      .setCustomId('signUp')
-      .setLabel('Sign Up')
-      .setStyle(ButtonStyle.Primary);
-  
-    const removeButton = new ButtonBuilder()
-      .setCustomId('remove')
-      .setLabel('Remove')
-      .setStyle(ButtonStyle.Danger);
-  
-    const buttons = [signUpButton, removeButton];
-  
-    // Check if the user has the required role to see the edit button
-    const memberRoles = interaction.member?.roles;
-    if (
-      memberRoles instanceof GuildMemberRoleManager && 
-      memberRoles.cache.has(process.env.MANAGER_ROLE_ID as string)
-    ) {
-      const editButton = new ButtonBuilder()
-        .setCustomId('edit')
-        .setLabel('Edit')
-        .setStyle(ButtonStyle.Secondary);
-      buttons.push(editButton);
-    }
-  
-    const row = new ActionRowBuilder<ButtonBuilder>().addComponents(buttons);
-  
-    if (interaction.customId === 'createEventModal') {
-      await interaction.reply({ content: template, components: [row] });
-    } else if (interaction.customId === 'editEventModal') {
-      await interaction.deferUpdate();
-      await interaction.editReply({ content: template, components: [row] });
-    }
-  };
-
