+++ src/interactions/utils/removeEvent.ts
