+++ src/interactions/utils/signupEvent.ts
