+++ src/interactions/utils/templates/eventTemplate.ts
