+++ src/interactions/utils/templates/moonlitTemplate.ts
@@ -1,0 +1,21 @@
+export function getMoonlitTemplate(
+    startTimeUnix: string,
+    dj: string,
+    leader: string,
+    slotList: string,
+  ) {
+    const template = `**LAYER.// Moonlit EU/NA**
+
+**Notes:**
+- The event starts on <t:${startTimeUnix}:F>.
+- There are 4 slots with 5 people per slot.
+${leader ? `- The leader of this event is ${leader}.` : ''}
+- Each slot lasts 15 minutes.
+
+**Slot Times:**
+[${dj}]
+${slotList}
+`.trim();
+  
+    return template;
+  }
